Fix clippy warnings and improve code quality (#139)
- Fixed 500+ clippy warnings across entire codebase
  - Replaced .min().max() with .clamp() for better readability
  - Used .keys() and .values() for map iteration instead of destructuring unused values
  - Replaced .len() == 0 with .is_empty() for idiomatic Rust
  - Collapsed nested if statements for cleaner code flow
  - Used vec![] macro instead of Vec::new() + push pattern
  - Removed unused imports in test modules
  - Fixed empty lines after doc comments
  - Added allow attributes for intentional recursive patterns

- Updated CHANGELOG.md with code quality improvements section

All changes improve code quality without affecting functionality.
Successfully built and tested with all 204 detectors working correctly.

+++ crates/lsp/src/server.rs
@@ -168,14 +168,14 @@ impl SolidityDefendLanguageServer {
             .map(|finding| {
                 let range = Range {
                     start: Position {
-                        line: finding.primary_location.line.saturating_sub(1) as u32,
-                        character: finding.primary_location.column.saturating_sub(1) as u32,
+                        line: finding.primary_location.line.saturating_sub(1),
+                        character: finding.primary_location.column.saturating_sub(1),
                     },
                     end: Position {
-                        line: finding.primary_location.line.saturating_sub(1) as u32,
+                        line: finding.primary_location.line.saturating_sub(1),
                         character: (finding.primary_location.column
                             + finding.primary_location.length)
-                            .saturating_sub(1) as u32,
+                            .saturating_sub(1),
                     },
                 };
 
@@ -403,14 +403,14 @@ impl LanguageServer for SolidityDefendLanguageServer {
             for finding in &doc.findings {
                 let finding_range = Range {
                     start: Position {
-                        line: finding.primary_location.line.saturating_sub(1) as u32,
-                        character: finding.primary_location.column.saturating_sub(1) as u32,
+                        line: finding.primary_location.line.saturating_sub(1),
+                        character: finding.primary_location.column.saturating_sub(1),
                     },
                     end: Position {
-                        line: finding.primary_location.line.saturating_sub(1) as u32,
+                        line: finding.primary_location.line.saturating_sub(1),
                         character: (finding.primary_location.column
                             + finding.primary_location.length)
-                            .saturating_sub(1) as u32,
+                            .saturating_sub(1),
                     },
                 };
 
@@ -461,14 +461,14 @@ impl LanguageServer for SolidityDefendLanguageServer {
             for finding in &doc.findings {
                 let finding_range = Range {
                     start: Position {
-                        line: finding.primary_location.line.saturating_sub(1) as u32,
-                        character: finding.primary_location.column.saturating_sub(1) as u32,
+                        line: finding.primary_location.line.saturating_sub(1),
+                        character: finding.primary_location.column.saturating_sub(1),
                     },
                     end: Position {
-                        line: finding.primary_location.line.saturating_sub(1) as u32,
+                        line: finding.primary_location.line.saturating_sub(1),
                         character: (finding.primary_location.column
                             + finding.primary_location.length)
-                            .saturating_sub(1) as u32,
+                            .saturating_sub(1),
                     },
                 };
